# openapi.yaml
openapi: 3.0.3
info:
  title: HILLink API
  version: 1.0.0
  description: |
    Production-ready OpenAPI starter generated from current Next.js route handlers.
    Authentication is session/cookie-based (Supabase). Admin routes require admin role.
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: http://localhost:3001
    description: Local test
security:
  - cookieAuth: []
...
paths:
...
  /api/admin/delete-user:
    post:
      tags: [Admin]
      summary: Delete user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteUserRequest'
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/admin/athlete-verification:
    post:
      tags: [Admin]
      summary: Set athlete verification status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AthleteVerificationRequest'
      responses:
        '200':
          description: Verification updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AthleteVerificationResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/admin/business-access-tier:
    post:
      tags: [Admin]
      summary: Set business access tier override
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessAccessTierRequest'
      responses:
        '200':
          description: Access tier updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessAccessTierResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/athlete/withdraw-application:
    post:
      tags: [Athlete]
      summary: Drop campaign/application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawApplicationRequest'
      responses:
        '200':
          description: Application withdrawn/dropped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/athlete/sync-xp-challenges:
    post:
      tags: [Athlete]
      summary: Sync challenge XP events
      responses:
        '200':
          description: Sync complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncXpChallengesResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
  schemas:
    Uuid:
      type: string
      format: uuid
      example: 7658ce51-2694-425b-8cfb-8f5cc93cc6cb
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
...
    DeleteUserRequest:
      type: object
      required: [userId]
      properties:
        userId:
          $ref: '#/components/schemas/Uuid'
    AthleteVerificationRequest:
      type: object
      required: [userId, status]
      properties:
        userId:
          $ref: '#/components/schemas/Uuid'
        status:
          type: string
          enum: [pending, approved, rejected]
        reason:
          type: string
    AthleteVerificationResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        emailWarning:
          type: string
          nullable: true
...
    BusinessAccessTierRequest:
      type: object
      required: [userId]
      properties:
        userId:
          $ref: '#/components/schemas/Uuid'
        accessTier:
          type: string
          nullable: true
          enum: [starter, growth, scale, domination]
    BusinessAccessTierResponse:
      type: object
      required: [success, userId, paidTier, effectiveTier]
      properties:
        success:
          type: boolean
        userId:
          $ref: '#/components/schemas/Uuid'
        paidTier:
          type: string
        accessTierOverride:
          type: string
          nullable: true
        effectiveTier:
          type: string
        warning:
          type: string
          nullable: true
...
    WithdrawApplicationRequest:
      type: object
      required: [applicationId]
      properties:
        applicationId:
          $ref: '#/components/schemas/Uuid'
...
    SyncXpChallengesResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        granted:
          type: integer
        grantedTitles:
          type: array
          items:
            type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Not signed in
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient privileges
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'